Time: 0 ms (100%), Space: 8.6 MB (48.53%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -18,7 +18,7 @@ public:
             unsigned int reversed = 0;
             while (x > 0)
             {   
-                int digit = x % 10;
+                short int digit = x % 10;
                 reversed = reversed * 10 + digit;
                 x /= 10;
             }
